mlops_manager/templates/app/__APP_NAME__/requirements.txt
PyYAML
google_cloud_pipeline_components 
kfp
google-cloud-aiplatform==1.1.1